FastME-TaxAdd_BalME/requirements.txt
appnope==0.1.0
backcall==0.2.0
biopython==1.78
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
hdmedians==0.13
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
kiwisolver==1.2.0
lockfile==0.12.2
matplotlib==3.3.1
msgpack==1.0.0
natsort==7.0.1
numpy==1.19.2
pandas==1.1.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-bio==0.5.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
